+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.02.03`.
+ * Upgraded stable Flutter analysis SDK to `2.10.0`.
+ * Upgraded preview Flutter analysis SDK to `2.10.0`.
 
 ## `20220202t122400-all`
  * Bumped runtimeVersion to `2022.01.31`.